requirements.txt
requests 
beautifulsoup4
lxml
python-dotenv
ollama
scikit-learn
numpy
matplotlib
